third_party/blink/web_tests/platform/mac-mac10.14/external/wpt/css/selectors/invalidation/target-pseudo-in-has-expected.txt

This is a testharness.js-based test.
FAIL CSS Selectors Invalidation: target pseudo in :has() argument assert_equals: parent should be yellowgreen on fragment click expected "rgb(154, 205, 50)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
